feat(ui): Several updates for the GUI;
The GUI file has been updated. The old 'cubeviewer' file has been renamed to 'cubeviewer_old' to back up all the functionalities with Matplotlib. After the change, a new 'cubeviewer' file has been created with some PyQt5's functionalities. Also the configuration files has been updated.

Took 25 minutes
