+++ config/settings.py
@@ -1,1 +1,12 @@
+import platform
+
 """Global configurations for GUI"""
+
+# Window configuration
+WINDOW_WIDTH = 600
+WINDOW_HEIGHT = 300
+X_POSITION = 300
+Y_POSITION = 50
+
+DEBUG_MODE = True
+
